+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - error for completed game
 - verify avoiding check
 
+### Modify
+- castling process
+
 ## [0.2.0] - 2019-01-13
 ### Modified
 - error handling for Position module